requirements.txt
clip
clip_anytorch
diffusers
Flask
filelock
Pillow
pytest
segment_anything
supervision
torch
opencv-python
supervision
matplotlib
tqdm
regex
ftfy
onnx
onnxruntime
pycocotools
opencv-contrib-python
Requests
streamlit
opencv-contrib-python
opencv-python
numpy